+++ package.json
@@ -1,34 +1,33 @@
 {
-  "name": "webpack",
+  "name": "webpack-setup",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack serve --open",
     "build": "webpack"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/HAMMAS-SALEEM/webpack.git"
-  },
   "keywords": [],
   "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/HAMMAS-SALEEM/webpack/issues"
-  },
-  "homepage": "https://github.com/HAMMAS-SALEEM/webpack#readme",
+  "license": "MIT",
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
     "css-loader": "^6.5.1",
-    "file-loader": "^6.2.0",
-    "html-loader": "^3.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-airbnb-base": "^14.2.1",
+    "eslint-plugin-import": "^2.25.4",
     "html-webpack-plugin": "^5.5.0",
-    "img-loader": "^4.0.0",
     "style-loader": "^3.3.1",
-    "svg-inline-loader": "^0.8.2",
-    "webpack": "^5.65.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.7.3"
+    "stylelint": "^13.13.1",
+    "stylelint-config-standard": "^21.0.0",
+    "stylelint-csstree-validator": "^1.9.0",
+    "stylelint-scss": "^3.21.0",
+    "webpack": "^5.38.1",
+    "webpack-cli": "^4.7.2",
+    "webpack-dev-server": "^4.7.2"
+  },
+  "dependencies": {
+    "lodash": "^4.17.21"
   }
 }
