+++ src/index.html
@@ -9,7 +9,7 @@
       crossorigin="anonymous"
     ></script>
     <title>Wbpack Exercise</title>
-<script defer src="index.bundle.js"></script></head>
+</head>
 
 <body>
     <section>
